Refactor: Add HashFactory class returning hash function objects
